+++ apps/api/package.json
@@ -26,6 +26,7 @@
   },
   "dependencies": {
     "@elysiajs/bearer": "^1.3.0",
+    "@elysiajs/cors": "^1.3.3",
     "@elysiajs/jwt": "^1.3.2",
     "@elysiajs/swagger": "^1.3.1",
     "@hebo/db": "workspace:*",
